nav-fix - removed redundant hook, fixed initial nav handler
